+++ sstmac/hardware/packet_flow/packet_allocator.h
@@ -12,7 +12,7 @@ class packet_allocator : public sprockit::factory_type
 {
  public:
   virtual packet_flow_payload*
-  new_packet(long bytes, long byte_offset, message* msg) = 0;
+  new_packet(int bytes, long byte_offset, message* msg) = 0;
 };
 
 DeclareFactory(packet_allocator)
